+++ resources/js/components/copyright.jsx
