+++ resources/js/pages/Admin/MenuAspekPenilaian.jsx
@@ -2,7 +2,7 @@ import React from "react";
 // 👇 [UBAH] Impor hook yang diperlukan dari Inertia
 import { usePage, Link, router } from "@inertiajs/react";
 import { Trash2, Home, Pencil, Search } from "lucide-react";
-import Sidebar from "../../Components_Riko/Sidebar.js";
+import Sidebar from "../../components/Sidebar.js";
 
 
 export default function MenuAspekPenilaian() {
